+++ lib/ui/explore/ExplorePanel.dart
@@ -534,9 +534,12 @@ class ExplorePanelState extends State<ExplorePanel>
     }
     if (_shouldLoadGames(categories)) {
       List<DateTime?> gamesTimeFrame = _getGamesTimeFrame(eventFilter);
-      List<Explore>? games = await Sports().loadGames(startDate: gamesTimeFrame.first, endDate: gamesTimeFrame.last);
+      List<Game>? games = await Sports().loadGames(startDate: gamesTimeFrame.first, endDate: gamesTimeFrame.last);
       if (CollectionUtils.isNotEmpty(games)) {
-        explores.addAll(games!);
+        List<Game>? displayGames = _buildDisplayGames(games!);
+        if (CollectionUtils.isNotEmpty(displayGames)) {
+          explores.addAll(displayGames!);
+        }
       }
     }
     _sortExplores(explores);
@@ -568,7 +571,7 @@ class ExplorePanelState extends State<ExplorePanel>
       case EventsDisplayType.multiple:
         displayEvents = [];
         for (Event event in allEvents) {
-          if (event.isComposite) {
+          if (event.isMultiEvent) {
             displayEvents.add(event);
           }
         }
@@ -576,13 +579,39 @@ class ExplorePanelState extends State<ExplorePanel>
       case EventsDisplayType.single:
         displayEvents = [];
         for (Event event in allEvents) {
-          if (!event.isComposite) {
+          if (!event.isMultiEvent) {
             displayEvents.add(event);
           }
         }
         break;
     }
     return displayEvents;
+  }
+
+  List<Game>? _buildDisplayGames(List<Game> allGames) {
+    List<Game>? displayGames;
+    switch (_selectedEventsDisplayType) {
+      case EventsDisplayType.all:
+        displayGames = allGames;
+        break;
+      case EventsDisplayType.multiple:
+        displayGames = [];
+        for (Game game in allGames) {
+          if (game.isMoreThanOneDay) {
+            displayGames.add(game);
+          }
+        }
+        break;
+      case EventsDisplayType.single:
+        displayGames = [];
+        for (Game game in allGames) {
+          if (!game.isMoreThanOneDay) {
+            displayGames.add(game);
+          }
+        }
+        break;
+    }
+    return displayGames;
   }
 
   ///